+++ app/src/examples/RuntimeTests/ReanimatedRuntimeTestsRunner/TestRunner.ts
@@ -1,0 +1,147 @@
+import { useRef } from "react";
+import { TestCase, TestSuite } from "./types";
+import { TestComponent } from "./TestComponent";
+import { render, stopRecordingAnimationUpdates, unmockAnimationTimer } from "./RuntimeTestsApi";
+
+export class TestRunner {
+  private testSuites: TestSuite[] = [];
+  private currentTestSuite: TestSuite | null = null;
+  private currentTestCase: TestCase | null = null;
+  private renderHook: (component: any) => void = () => {};
+  private renderLock: { lock: boolean } = { lock: false };
+
+  public configure({ render }: { render: (component: any) => void }) {
+    this.renderHook = render;
+  }
+
+  public render(component: any) {
+    this.renderHook(component);
+  }
+
+  public getRenderLock() {
+    return this.renderLock;
+  }
+
+  public describe(name: string, buildSuite: () => void) {
+    this.testSuites.push({
+      name,
+      buildSuite,
+      testCases: [],
+    });
+  };
+
+  public test(name: string, testCase: () => void) {
+    if (!this.currentTestSuite) {
+      throw new Error("Undefined test suite context");
+    }
+    this.currentTestSuite.testCases.push({
+      name,
+      testCase,
+      componentsRefs: {},
+      errors: [],
+    });
+  };
+
+  public useTestRef(name: string): React.MutableRefObject<any> {
+    const ref = useRef();
+    if (!this.currentTestCase) {
+      throw new Error("Undefined test case context");
+    }
+    this.currentTestCase.componentsRefs[name] = ref;
+    return ref;
+  }
+
+  public getTestComponent(name: string): TestComponent {
+    if (!this.currentTestCase) {
+      throw new Error("Undefined test case context");
+    }
+    return new TestComponent(this.currentTestCase.componentsRefs[name]);
+  }
+
+  public async runTests() {
+    for (const testSuite of this.testSuites) {
+      this.currentTestSuite = testSuite;
+      console.log(`+ Running test suite: ${testSuite.name}`);
+      console.log(testSuite)
+      testSuite.buildSuite();
+      if (testSuite.beforeAll) {
+        await testSuite.beforeAll();
+      }
+      for (const testCase of testSuite.testCases) {
+        this.currentTestCase = testCase;
+        console.log(`\t - Running test case: ${testCase.name}`);
+        if (testSuite.beforeEach) {
+          await testSuite.beforeEach();
+        }
+        await testCase.testCase();
+        if (testCase.errors.length > 0) {
+          console.error(testCase.errors);
+          console.log("\t -----------------------------------------------");
+        } else {
+          console.log("\t ✅ OK");
+          console.log("\t -----------------------------------------------");
+        }
+        if (testSuite.afterEach) {
+          await testSuite.afterEach();
+        }
+        this.currentTestCase = null;
+        await render(null);
+      }
+      if (testSuite.afterAll) {
+        await testSuite.afterAll();
+      }
+      this.currentTestSuite = null;
+    }
+    this.testSuites = [];
+    await unmockAnimationTimer();
+    await stopRecordingAnimationUpdates();
+    console.log("End of tests run 🏁");
+  }
+
+  public expect(value: any) {
+    if (!this.currentTestCase) {
+      throw new Error("Undefined test case context");
+    }
+    const errors = this.currentTestCase?.errors;
+    return {
+      toBe: (expected: any) => {
+        if (value !== expected) {
+          errors.push(`Expected ${value} received ${expected}`);
+        }
+      },
+      toMatchSnapshot: (expected: any) => {
+        if (JSON.stringify(value) !== JSON.stringify(expected)) {
+          errors.push(`Expected ${JSON.stringify(value)} received ${JSON.stringify(expected)}`);
+        }
+      }
+    };
+  };
+
+  public beforeAll(job: () => void) {
+    if (!this.currentTestSuite) {
+      throw new Error("Undefined test suite context");
+    }
+    this.currentTestSuite.beforeAll = job;
+  }
+
+  public afterAll(job: () => void) {
+    if (!this.currentTestSuite) {
+      throw new Error("Undefined test suite context");
+    }
+    this.currentTestSuite.afterAll = job;
+  }
+
+  public beforeEach(job: () => void) {
+    if (!this.currentTestSuite) {
+      throw new Error("Undefined test suite context");
+    }
+    this.currentTestSuite.beforeEach = job;
+  }
+
+  public afterEach(job: () => void) {
+    if (!this.currentTestSuite) {
+      throw new Error("Undefined test suite context");
+    }
+    this.currentTestSuite.afterEach = job;
+  }
+}
