+++ app/src/examples/RuntimeTests/RuntimeTestsExample.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import RuntimeTestsRunner from './ReanimatedRuntimeTestsRunner/RuntimeTestsRunner';
+
+// load tests
+import './tests/Animations.test';
+
+export default function RuntimeTestsExample() {
+  return (
+    <RuntimeTestsRunner />
+  );
+}
+
