+++ app/src/examples/RuntimeTests/tests/snapshots/Animations.snapshot.ts
@@ -1,0 +1,4 @@
+export const Snapshots = {
+  animation3: [{"width": 0}, {"width": 0.20479999999999998}, {"width": 0.8191999999999999}, {"width": 1.8432}, {"width": 3.2767999999999997}, {"width": 5.12}, {"width": 7.3728}, {"width": 10.035200000000001}, {"width": 13.107199999999999}, {"width": 16.5888}, {"width": 20.48}, {"width": 24.780799999999996}, {"width": 29.4912}, {"width": 34.6112}, {"width": 40.140800000000006}, {"width": 46.08}, {"width": 52.371199999999995}, {"width": 58.41280000000001}, {"width": 64.0448}, {"width": 69.2672}, {"width": 74.08}, {"width": 78.48320000000001}, {"width": 82.4768}, {"width": 86.0608}, {"width": 89.2352}, {"width": 92}, {"width": 94.3552}, {"width": 96.3008}, {"width": 97.8368}, {"width": 98.9632}, {"width": 99.68}, {"width": 99.9872}, {"width": 100}],
+  layoutAnimation: [{"opacity": 0}, {"opacity": 0.02275555555555556}, {"opacity": 0.09102222222222224}, {"opacity": 0.2048}, {"opacity": 0.36408888888888896}, {"opacity": 0.5644444444444444}, {"opacity": 0.7408}, {"opacity": 0.8716444444444444}, {"opacity": 0.9569777777777778}, {"opacity": 0.9968}, {"opacity": 1}],
+};
